+++ frontend/src/types/Bet.ts
@@ -1,0 +1,4 @@
+export interface Bet {
+  fixture: number;
+  team: string;
+}
